requirements.txt
python-telegram-bot==20.8
python-dotenv
fastapi
uvicorn
aiohttp
tinydb
matplotlib
yfinance
openai>=1.10.0
ta
scikit-learn
joblib
pyrogram
tgcrypto
nest_asyncio
binance
cryptography